Fixed the chrome bookmarklet
It "technically works". The best kind of "works" (or in this case, it's
a horrible hack that sends messages around everywhere, litters the code
with checks to see if we're in a chrome extension, and duplicates the
whole code base into the chrome-extension folder. Also, it uses popups,
which I didn't really want to do).

This means that you can now bookmark stuff on GitHub again.